pyproject.toml
[tool.poetry]
name = "api-commerce"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "api_commerce"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.2.2"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django = "^4.2"
dj-database-url = "^2.0.0"
whitenoise = "^6.4.0"
django-filter = "^23.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pycodestyle = "^2.10.0"
ipdb = "^0.13.13"
ipython = "^8.13.1"
drf-spectacular = "^0.26.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
